adding fixes
